passage en virgule fixe (debut)

+++ include/kernel/trapeze.h
@@ -5,18 +5,20 @@
 //! @brief Trapezoidal speed
 //! @author Atlantronic
 
+#include <stdint.h>
+
 struct trapeze
 {
-	float a_max;
-	float d_max;
-	float v_max;
+	int32_t a_max;
+	int32_t d_max;
+	int32_t v_max;
 
-	float s;
-	float v;
+	int32_t s;
+	int32_t v;
 };
 
-void trapeze_apply(struct trapeze* t, float s);
+void trapeze_apply(struct trapeze* t, int32_t s);
 
-void trapeze_reset(struct trapeze* t, float s, float v);
+void trapeze_reset(struct trapeze* t, int32_t s, int32_t v);
 
 #endif
